Move compiler passes to their own namespace
